% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datas.R
\docType{data}
\name{datas}
\alias{datas}
\title{An example for create a description of your data}
\format{
An object of class \code{"data.frame"}
\describe{
\item{X1}{Data generated from normal}
\item{X2}{Data generated from uniform}
\item{X3}{Data generated from normal}
}
}
\usage{
datas
}
\description{
Data set for testing data
}
\references{
This data set was artificially created for the overviewR package.
}
\keyword{datasets}
