% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnl.R
\name{tnl}
\alias{tnl}
\title{The (tnl) function.}
\usage{
tnl(k,n,l)
}
\arguments{
\item{k}{Vector of quantiles.}

\item{n}{Sample size.}

\item{l}{The class parameter of Tnl.}
}
\value{
list of an exact probability and distribution function.
}
\description{
Perform exact tnl.
}
\details{
vbcdfgdfg
}
\examples{
 ##You need to select all of \code{k}, \code{n} and \code{l}
tnl(k=2,n=7,l=1)
\dontrun{
$prob
[1] 0.1538462

$tt
[1] 0.2307692
}

}
\references{
A Nonparametric Test for the Two-Sample Problem Based on Order Statistics
}
